Added missing cast to long for status code

+++ warp10/src/main/java/io/warp10/script/ext/http/HTTP.java
@@ -1,5 +1,5 @@
 //
-//   Copyright 2021-2023  SenX S.A.S.
+//   Copyright 2021-2025  SenX S.A.S.
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -437,7 +437,7 @@ public class HTTP extends NamedWarpScriptFunction implements WarpScriptStackFunc
       // Form response
       //
 
-      res.put(STATUS_CODE, conn.getResponseCode());
+      res.put(STATUS_CODE, (long) conn.getResponseCode());
       Map<String, List<String>> hdrs = conn.getHeaderFields();
 
       // headers map is immutable
